requirements.txt
flask
transformers
newspaper3k
feedparser